move some stuff to merkle file
